ordem-de-servico/requirements.txt
django-extensions==3.2.3
django-localflavor==4.0
django-ninja==0.22.2
django-widget-tweaks==1.5.0
dj-database-url==2.3.0
Django==4.2.3
psycopg2-binary==2.9.10
python-decouple==3.8
isort==5.12.0
autopep8==2.0.2
djhtml==3.0.6
asgiref==3.8.1
pillow==11.1.0
pyaes==1.6.1
pyrogram-repl==2.0.106
PySocks==1.7.1
sqlparse==0.5.3
typing_extensions==4.12.2
